requirements.txt
hurry.filesize >= 0.9